+++ packages/backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @m5kdev/backend
 
+## 0.3.2
+
+### Patch Changes
+
+- table
+- Updated dependencies
+  - @m5kdev/commons@0.3.2
+  - @m5kdev/config@0.3.2
+
 ## 0.3.1
 
 ### Patch Changes