+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @m5kdev/config
 
+## 0.3.2
+
+### Patch Changes
+
+- table
+
 ## 0.3.1
 
 ### Patch Changes